tests: Common test for NPC entity
